test: TestContainersConfig 클래스로 TestContainer 설정 부분을 분리
